Title: Update version from 0.8.1 to 0.8.2
Description:
This commit updates the version of the cloudforge package from 0.8.1 to 0.8.2. The version change is reflected in the VERSION file, the __init__.py file, and the pyproject.toml file. This update is necessary to reflect the new changes, improvements, or bug fixes that have been implemented in the package.

Added:
No new features or files were added in this commit.

Removed:
No features or files were removed in this commit.

Modified:
- VERSION: Updated the version number from 0.8.1 to 0.8.2.
- cloudforge/__init__.py: Updated the __version__ variable from 0.8.1-dev to 0.8.2-dev.
- pyproject.toml: Updated the version number under [tool.poetry] from 0.8.1-dev to 0.8.2-dev.